+++ tests/client/branches_test.py
@@ -5,6 +5,7 @@ Tests for the Branches endpoint.
 import pytest
 
 PROJECT_ID = "454087345e09f3e7e7eae3.57891254"
+PROJECT_ID2 = "2273827860c1e2473eb195.11207948"
 BRANCH_ID = 55312
 NEW_BRANCH_ID = 66307
 
@@ -22,19 +23,18 @@ def test_all_branches(client):
 def test_all_branches_pagination(client):
     """Tests fetching of all branches with pagination params
     """
-    branches = client.branches(PROJECT_ID, {"page": 2, "limit": 1})
-    assert branches.project_id == PROJECT_ID
-    assert branches.items[0].branch_id == 45845
-    # revise these tests once ITSD-854 is resolved
+    branches = client.branches(PROJECT_ID2, {"page": 2, "limit": 1})
+    assert branches.project_id == PROJECT_ID2
+    assert branches.items[0].branch_id == 362586
     assert branches.current_page == 2
-    assert branches.total_count == 1
-    assert branches.page_count == 1
+    assert branches.total_count == 4
+    assert branches.page_count == 4
     assert branches.limit == 1
 
-    # assert branches.is_last_page()
-    # assert branches.is_first_page()
-    # assert not branches.has_next_page()
-    # assert not branches.has_prev_page()
+    assert not branches.is_last_page()
+    assert not branches.is_first_page()
+    assert branches.has_next_page()
+    assert branches.has_prev_page()
 
 
 @pytest.mark.vcr